+++ caddyhttp/requestid/setup.go
@@ -1,0 +1,27 @@
+package requestid
+
+import (
+	"github.com/mholt/caddy"
+	"github.com/mholt/caddy/caddyhttp/httpserver"
+)
+
+func init() {
+	caddy.RegisterPlugin("requestid", caddy.Plugin{
+		ServerType: "http",
+		Action:     setup,
+	})
+}
+
+func setup(c *caddy.Controller) error {
+	for c.Next() {
+		if c.NextArg() {
+			return c.ArgErr() //no arg expected.
+		}
+	}
+
+	httpserver.GetConfig(c).AddMiddleware(func(next httpserver.Handler) httpserver.Handler {
+		return Handler{Next: next}
+	})
+
+	return nil
+}
